requirements.txt
boto3==1.26.5
botocore==1.29.5
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
gTTS==2.2.4
idna==3.4
jmespath==1.0.1
mutagen==1.46.0
pip==21.2.3
pydub==0.25.1
python-dateutil==2.8.2
PyYAML==6.0
requests==2.28.1
s3transfer==0.6.0
setuptools==57.4.0
six==1.16.0
urllib3==1.26.12
